+++ example/lib/example/list_wheel_example.dart
@@ -71,7 +71,7 @@ class _ListWheelExampleState extends State<ListWheelExample> {
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-          _observer.debugCheckOnstageItems(
+          _observer.getVisibleItems(
             scrollExtent: ScrollExtent.fromPosition(_controller.position),
           );
         },